requirements.txt
openai
openpyxl
requests
pandas
beautifulsoup4
numpy<2
uvicorn
fastapi
rich
python-dotenv
fastapi-cors
chardet
pytrends
Pillow
datasets
transformers
sentence-transformers
pinecone
pinecone-text
protobuf==3.20.3
